Version 1.0.0

+++ lib/externals/repository.rb
@@ -3,27 +3,19 @@ require 'fileutils'
 module Externals
   class Repository
     
+    attr_reader :name
+    
     def initialize(base_dir, name, repo_url, rel_path)
       @base_dir = base_dir
       @name = name
       @repo_url = repo_url
       @rel_path = rel_path
     end
-    attr_reader :name
     
-    def install
-      # Create directory that we will clone into
-      unless File.exist?(checkout_path)
-        FileUtils.mkdir_p(checkout_path)
-      end
-      Dir.chdir(checkout_path) do
-        # Remove repository if exists
-        FileUtils.rm_rf(@name)
-        # Clone repository
-        `git clone #{@repo_url} #{@name}`
-      end
+    def exists?
+      File.exist?(repo_path)
     end
-
+    
     def freeze
       install unless exists?
       if is_not_a_git_repo?
@@ -40,13 +32,42 @@ module Externals
             # Make temp directory
             FileUtils.mkdir_p(temp_path)
             # Compress .git folder to temp
-            `tar czf #{temp_path}/#{@name}.git.tgz .git`
+            `tar czf #{temp_path}/#{@name}.git.tgz .git` unless $TESTING
           end
           # Remove repository's .git folder
           FileUtils.rm_r('.git')
         end
         puts "#{@name} frozen"
       end
+    end
+    
+    def install
+      # Create directory that we will clone into
+      unless File.exist?(checkout_path)
+        FileUtils.mkdir_p(checkout_path)
+      end
+      Dir.chdir(checkout_path) do
+        # Remove repository if exists
+        FileUtils.rm_rf(@name)
+        # Clone repository
+        `git clone #{@repo_url} #{@name}` unless $TESTING
+      end
+    end
+    
+    def is_a_git_repo?
+      File.exist?("#{repo_path}/.git")
+    end
+    
+    def is_not_a_git_repo?
+      !is_a_git_repo?
+    end
+    
+    def is_compressed?
+      File.exists?("#{temp_path}/#{@name}.git.tgz")
+    end
+    
+    def is_not_compressed?
+      !is_compressed?
     end
     
     def status
@@ -67,7 +88,7 @@ module Externals
         if is_compressed?
           Dir.chdir(temp_path) do
             # Decompress git snapshot
-            `tar xzf #{@name}.git.tgz`
+            `tar xzf #{@name}.git.tgz` unless $TESTING
             # Move back to repo
             FileUtils.mv(".git", repo_path)
             # Remove snapshot
@@ -75,7 +96,7 @@ module Externals
           end
           # Reset repository to snapshot
           Dir.chdir(repo_path) do
-            `git reset --hard`
+            `git reset --hard` unless $TESTING
           end
         else
           # Clone fresh repo if no snapshot found
@@ -84,28 +105,9 @@ module Externals
         puts "#{@name} unfrozen"
       end
     end
-    
-    def exists?
-      File.exist?(repo_path)
-    end
-    
-    def is_a_git_repo?
-      File.exist?("#{repo_path}/.git")
-    end
-    
-    def is_not_a_git_repo?
-      !is_a_git_repo?
-    end
-    
-    def is_compressed?
-      File.exists?("#{temp_path}/#{@name}.git.tgz")
-    end
-    
-    def is_not_compressed?
-      !is_compressed?
-    end
 
     private
+    
     def checkout_path
       File.expand_path(File.join(@base_dir, @rel_path))
     end
